EISW-163565 [Cleanup] Revert "Expose TimingManager OutputTextStrategy"
This reverts commit 7ce262291007d7f6961cefa7c41a809abd5bf30a.
